Server: arrange directories
